ZooKeeperConfiguration으로 경로 설정 이동
