+++ config0.yaml
@@ -6,9 +6,9 @@ client:
 cluster:
   address: 'localhost'
   port: 9090
-  path: 'databuter'
 zooKeeper:
   address: 'localhost'
   port: 2181
+  path: 'databuter'
   baseSleepTimeMs: 1000
   maxRetries: 8
